Add FLUX.2 TAESD preview skip and VAE dtype fix
FLUX.2 uses 32-channel latents which have no compatible TAESD decoder.
Skip preview generation for FLUX.2 models instead of failing.

Changes:
- Add is_flux2 parameter to TAESDManager.load_taesd() and decode_latent()
- Return None for FLUX.2 models (no preview available)
- Add is_flux2 detection in routes.py for all generation endpoints
- Update create_progress_callback_factory with is_flux2 parameter
- Fix VAE dtype mismatch by converting latents before decode

+++ backend/core/pipeline.py
@@ -1767,7 +1767,8 @@ class DiffusionPipelineManager:
             # Unpatchify
             latents = self._flux2_unpatchify_latents(latents)
 
-            # Decode
+            # Decode - convert latents to VAE dtype (bfloat16 -> float32)
+            latents = latents.to(dtype=vae.dtype)
             with torch.no_grad():
                 image = vae.decode(latents, return_dict=False)[0]
